+++ src/tno/euphorie/report.py
@@ -304,8 +304,6 @@ class TNOMeasuresOverview(MeasuresOverview):
                  'classes': classes,
                  })
 
-        # re-use top-level module computation from the Status overview
-        modules = self.getModules()
         main_modules = {}
         for module, risks in sorted(modulesdict.items(), key=lambda m: m[0].zodb_path):
             module_obj = self.request.survey.restrictedTraverse(module.zodb_path.split('/'))
@@ -324,8 +322,8 @@ class TNOMeasuresOverview(MeasuresOverview):
                     else:
                         main_modules[path]['risks'][prio] = risks[prio]
             else:
-                title = modules[path]['title']
-                number = modules[path]['number']
+                title = module.title
+                number = module.number
                 main_modules[path] = {'name': title, 'number': number, 'risks': risks}
 
         self.modules = []